Struct matrix_sdk_crypto_nodejs::encryption::EncryptionSettings

pub struct EncryptionSettings {
    pub algorithm: EncryptionAlgorithm,
    pub rotation_period: BigInt,
    pub rotation_period_messages: BigInt,
    pub history_visibility: HistoryVisibility,
}

Settings for an encrypted room.

This determines the algorithm and rotation periods of a group session.

Fields
algorithm: EncryptionAlgorithm

The encryption algorithm that should be used in the room.

rotation_period: BigInt

How long the session should be used before changing it, expressed in microseconds.

rotation_period_messages: BigInt

How many messages should be sent before changing the session.

history_visibility: HistoryVisibility

The history visibility of the room when the session was created.

Implementations

impl EncryptionSettings

pub fn into_reference(
val: EncryptionSettings,
env: Env
) -> Result<Reference<EncryptionSettings>>

pub fn into_instance(self, env: Env) -> Result<ClassInstance<EncryptionSettings>>

impl EncryptionSettings

pub fn new() -> EncryptionSettings

Create a new EncryptionSettings with default values.