Struct matrix_sdk_crypto_nodejs::machine::OlmMachine

pub struct OlmMachine { /* private fields */ }

State machine implementation of the Olm/Megolm encryption protocol used for Matrix end to end encryption.

Implementations

impl OlmMachine

pub fn into_reference(val: OlmMachine, env: Env) -> Result<Reference<OlmMachine>>

pub fn into_instance(self, env: Env) -> Result<ClassInstance<OlmMachine>>

impl OlmMachine

pub async fn initialize(
user_id: &UserId,
device_id: &DeviceId,
store_path: Option<String>,
store_passphrase: Option<String>
) -> Result<OlmMachine>

Create a new memory-based OlmMachine asynchronously.

The persistence of the encryption keys and all the inner objects are controlled by the store_path argument.

Arguments

user_id, the unique ID of the user that owns this machine.

device_id, the unique id of the device that owns this machine.

store_path, the path to a directory where the state of the machine should be persisted; if not set, the created machine will keep the encryption keys only in memory, and once the object is dropped, the keys will be lost.

store_passphrase, the passphrase that should be used to encrypt the data at rest in the store. Warning, if no passphrase is given, the store and all its data will remain unencrypted. This argument is ignored if store_path is not set.

pub fn new() -> Result<()>

It’s not possible to construct an OlmMachine with its constructor because building an OlmMachine is asynchronous. Please use the finalize method.

pub fn user_id(&self) -> UserId

The unique user ID that owns this OlmMachine instance.

pub fn device_id(&self) -> DeviceId

The unique device ID that identifies this OlmMachine.

pub fn identity_keys(&self) -> IdentityKeys

Get the public parts of our Olm identity keys.

pub async fn receive_sync_changes(
&self,
to_device_events: String,
changed_devices: &DeviceLists,
one_time_key_counts: HashMap<String, u32>,
unused_fallback_keys: Vec<String>
) -> Result<String>

Handle a to-device and one-time key counts from a sync response.

This will decrypt and handle to-device events returning the decrypted versions of them, as a JSON-encoded string.

To decrypt an event from the room timeline, please use decrypt_room_event.

Arguments

to_device_events, thhe to-device events of the current sync response.

changed_devices, the list of devices that changed in this sync response.

one_time_keys_count, the current one-time keys counts that the sync response returned.

pub async fn outgoing_requests(
&self
) -> Result<Vec<Either7<KeysUploadRequest, KeysQueryRequest, KeysClaimRequest, ToDeviceRequest, SignatureUploadRequest, RoomMessageRequest, KeysBackupRequest>>>

Get the outgoing requests that need to be sent out.

This returns a list of KeysUploadRequest, or KeysQueryRequest, or KeysClaimRequest, or ToDeviceRequest, or SignatureUploadRequest, or RoomMessageRequest, or KeysBackupRequest. Those requests need to be sent out to the server and the responses need to be passed back to the state machine using mark_request_as_sent.

pub async fn mark_request_as_sent(
&self,
request_id: String,
request_type: RequestType,
response: String
) -> Result<bool>

Mark the request with the given request ID as sent.

Arguments

request_id, the unique ID of the request that was sent out. This is needed to couple the response with the now sent out request.

request_type, the request type associated to the request ID.

response, the response that was received from the server after the outgoing request was sent out.

pub async fn get_missing_sessions(
&self,
users: Option<Vec<&UserId>>
) -> Result<Option<KeysClaimRequest>>

Get the a key claiming request for the user/device pairs that we are missing Olm sessions for.

Returns null if no key claiming request needs to be sent out.

Sessions need to be established between devices so group sessions for a room can be shared with them.

This should be called every time a group session needs to be shared as well as between sync calls. After a sync some devices may request room keys without us having a valid Olm session with them, making it impossible to server the room key request, thus it’s necessary to check for missing sessions between sync as well.

Note: Care should be taken that only one such request at a time is in flight, e.g. using a lock.

The response of a successful key claiming requests needs to be passed to the OlmMachine with the mark_request_as_sent.

Arguments

users, the list of users that we should check if we lack a session with one of their devices. This can be an empty array or null when calling this method between sync requests.

pub async fn update_tracked_users(&self, users: Vec<&UserId>)

Update the tracked users.

This will mark users that weren’t seen before for a key query and tracking.

If the user is already known to the Olm machine it will not be considered for a key query.

Arguments

users, an array over user IDs that should be marked for tracking.

pub async fn share_room_key(
&self,
room_id: &RoomId,
users: Vec<&UserId>,
encryption_settings: &EncryptionSettings
) -> Result<String>

Get to-device requests to share a room key with users in a room.

Arguments

room_id, the room ID of the room where the room key will be used.

users, the list of users that should receive the room key.

encryption_settings, the encryption settings.

pub async fn encrypt_room_event(
&self,
room_id: &RoomId,
event_type: String,
content: String
) -> Result<String>

Encrypt a JSON-encoded content for the given room.

Arguments

room_id, the ID of the room for which the message should be encrypted.

event_type, the plaintext type of the event.

content, the JSON-encoded content of the message that should be encrypted.

pub async fn decrypt_room_event(
&self,
event: String,
room_id: &RoomId
) -> Result<DecryptedRoomEvent>

Decrypt an event from a room timeline.

Arguments

event, the event that should be decrypted.

room_id, the ID of the room where the event was sent to.

pub async fn cross_signing_status(&self) -> CrossSigningStatus

Get the status of the private cross signing keys.

This can be used to check which private cross signing keys we have stored locally.

pub async fn sign(&self, message: String) -> Signatures

Sign the given message using our device key and if available cross-signing master key.