Module matrix_sdk_crypto_nodejs::requests

Types to handle requests.

Structs

KeysBackupRequest

A request that will back up a batch of room keys to the server (specification).

KeysClaimRequest

Data for a request to the /keys/claim API endpoint (specification).

KeysQueryRequest

Data for a request to the /keys/query API endpoint (specification).

KeysUploadRequest

Data for a request to the /keys/upload API endpoint (specification).

RoomMessageRequest

A customized owned request type for sending out room messages (specification).

SignatureUploadRequest

Data for a request to the /keys/signatures/upload API endpoint (specification).

ToDeviceRequest

Data for a request to the /sendToDevice API endpoint (specification).

Enums

RequestType

Represent the type of a request.

Type Definitions

OutgoingRequests