Struct matrix_sdk_crypto_nodejs::types::Signatures

pub struct Signatures { /* private fields */ }

Implementations

impl Signatures

pub fn into_reference(val: Signatures, env: Env) -> Result<Reference<Signatures>>

pub fn into_instance(self, env: Env) -> Result<ClassInstance<Signatures>>

impl Signatures

pub fn new() -> Self

Creates a new, empty, signatures collection.

pub fn add_signature(
&mut self,
signer: &UserId,
key_id: &DeviceKeyId,
signature: &Ed25519Signature
) -> Option<MaybeSignature>

Add the given signature from the given signer and the given key ID to the collection.

pub fn get_signature(
&self,
signer: &UserId,
key_id: &DeviceKeyId
) -> Option<Ed25519Signature>

Try to find an Ed25519 signature from the given signer with the given key ID.

pub fn get(&self, signer: &UserId) -> Option<HashMap<String, MaybeSignature>>

Get the map of signatures that belong to the given user.

pub fn clear(&mut self)

Remove all the signatures we currently hold.

pub fn is_empty(&self) -> bool

Do we hold any signatures or is our collection completely empty.

pub fn count(&self) -> usize

How many signatures do we currently hold.